Added a MainMenu
Here I added a Main Menu as a separate folder to be implemented into the game.

It has a START button to start the level.
HELP opens up an instructions window.
AUDIO opens up a window with a slider to change the Master Volume.
EXIT stops the game.
Each subwindow also has a BACK button.

The code for navigation includes IF statements.